Fix root files links have an invalid href
When rendering files in the root directory, an inadvertent `/` was being
prepended to the file path that is rendered as the href value. This
issue was introduced in ca980550af95fde09e803d4035051f435db85011. We
would always prepend a `/` character to the persisted path. In the case
of an empty persisted path, the current path would still be set to `/`
instead of a blank string as expected.